+++ classes/class-setup.php
@@ -100,7 +100,7 @@ class Setup {
 		/**
 		 * Load Gutenberg Pullquote Block Styles
 		 */
-		require_once LSX_BLOCKS_PATH . 'src/extensions/block-pullquote/index.php';
+		//require_once LSX_BLOCKS_PATH . 'src/extensions/block-pullquote/index.php';
 
 		/**
 		 * Load Team Block